fix: SmartReleaseScheduler episodeNumber Long→Double

+++ ANI-KUTA/APP/ani-kuta/core/updates/src/main/java/com/confused/anikuta/core/updates/SmartReleaseScheduler.kt
@@ -57,7 +57,7 @@ class SmartReleaseScheduler(
             SmartReleaseCheckWorker.schedule(
                 context = context,
                 mainId = state.mainId,
-                episodeNumber = episodeNumber,
+                episodeNumber = episodeNumber.toDouble(),
                 airingAt = airingAt,
                 attempt = 1,
             )
